+++ .gitignore
@@ -1,45 +1,3 @@
-/[Ll]ibrary/
-/[Tt]emp/
-/[Oo]bj/
-/[Bb]uild/
-/[Bb]uilds/
-/[Ll]ogs/
+.idea/**
+.vs/**
 
-/Assets/AssetStoreTools*
-/Assets/Plugins*
-/Assets/ProjectSettings*
-
-# Visual Studio 2015 cache directory
-/.vs/
-
-# Autogenerated VS/MD/Consulo solution and project files
-ExportedObj/
-.consulo/
-*.csproj
-*.unityproj
-*.sln
-*.suo
-*.tmp
-*.user
-*.userprefs
-*.pidb
-*.booproj
-*.svd
-*.pdb
-
-
-# Unity3D generated meta files
-*.pidb.meta
-
-# Unity3D Generated File On Crash Reports
-sysinfo.txt
-
-# Builds
-*.apk
-*.unitypackage
-
-# Visual Studio Code directory
-.vscode/
-
-# Editor user settings
-/[Uu]serSettings/EditorUserSettings.asset
