--- Assets/Editor.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f2f87ae8e11034977bd131ccc24a68e6
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
