--- Assets/Editor/HistoryNavigator.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4e826b6e6e507401e943369f3c1d41fd
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
