--- Assets/Editor/HistoryNavigator/CGTK.Tools.HistoryNavigator.asmdef
@@ -1,16 +1,0 @@
-{
-    "name": "CGTK.Tools.HistoryNavigator",
-    "rootNamespace": "CGTK.Tools.HistoryNavigator",
-    "references": [],
-    "includePlatforms": [
-        "Editor"
-    ],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": false,
-    "defineConstraints": [],
-    "versionDefines": [],
-    "noEngineReferences": false
-}
