--- Assets/Editor/HistoryNavigator/HistoryNavigatorOverlay.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2e61c229059774013b4f854e91abebb2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
