--- Assets/Editor/HistoryNavigator/Resources.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4585be5fdd6e84f5c8236a5d07956ed7
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
