--- Assets/Editor/HistoryNavigator/RightArrowButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 60b5408c858684493802c7920da10f44
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
