--- Assets/Scenes.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6ea315d0fd7389c41b19996891e99ae3
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
