+++ Editor/CGTK.Tools.HistoryNavigator.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 33c22d63313981d4fabf500d7288d6c5
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
