+++ Editor/LeftArrowButton.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 18f1cde5168964449a74c48ece648f43
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
