--- Logs/Packages-Update.log
@@ -1,46 +1,0 @@
-
-=== Wed Dec 29 18:59:07 2021
-
-Packages were changed.
-Update Mode: mergeDefaultDependencies
-
-The following packages were added:
-  com.unity.ide.rider@3.0.7
-  com.unity.ide.visualstudio@2.0.12
-  com.unity.ide.vscode@1.2.4
-  com.unity.modules.ai@1.0.0
-  com.unity.modules.androidjni@1.0.0
-  com.unity.modules.animation@1.0.0
-  com.unity.modules.assetbundle@1.0.0
-  com.unity.modules.audio@1.0.0
-  com.unity.modules.cloth@1.0.0
-  com.unity.modules.director@1.0.0
-  com.unity.modules.imageconversion@1.0.0
-  com.unity.modules.imgui@1.0.0
-  com.unity.modules.jsonserialize@1.0.0
-  com.unity.modules.particlesystem@1.0.0
-  com.unity.modules.physics@1.0.0
-  com.unity.modules.physics2d@1.0.0
-  com.unity.modules.screencapture@1.0.0
-  com.unity.modules.terrain@1.0.0
-  com.unity.modules.terrainphysics@1.0.0
-  com.unity.modules.tilemap@1.0.0
-  com.unity.modules.ui@1.0.0
-  com.unity.modules.uielements@1.0.0
-  com.unity.modules.umbra@1.0.0
-  com.unity.modules.unityanalytics@1.0.0
-  com.unity.modules.unitywebrequest@1.0.0
-  com.unity.modules.unitywebrequestassetbundle@1.0.0
-  com.unity.modules.unitywebrequestaudio@1.0.0
-  com.unity.modules.unitywebrequesttexture@1.0.0
-  com.unity.modules.unitywebrequestwww@1.0.0
-  com.unity.modules.vehicles@1.0.0
-  com.unity.modules.video@1.0.0
-  com.unity.modules.vr@1.0.0
-  com.unity.modules.wind@1.0.0
-  com.unity.modules.xr@1.0.0
-  com.unity.test-framework@1.1.29
-  com.unity.visualscripting@1.7.6
-The following packages were updated:
-  com.unity.collab-proxy from version 1.13.5 to 1.15.4
-  com.unity.timeline from version 1.6.2 to 1.6.3
