--- ProjectSettings/PackageManagerSettings.asset
@@ -1,35 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &1
-MonoBehaviour:
-  m_ObjectHideFlags: 61
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 13964, guid: 0000000000000000e000000000000000, type: 0}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  m_EnablePreReleasePackages: 0
-  m_EnablePackageDependencies: 0
-  m_AdvancedSettingsExpanded: 1
-  m_ScopedRegistriesSettingsExpanded: 1
-  m_SeeAllPackageVersions: 0
-  oneTimeWarningShown: 0
-  m_Registries:
-  - m_Id: main
-    m_Name: 
-    m_Url: https://packages.unity.com
-    m_Scopes: []
-    m_IsDefault: 1
-    m_Capabilities: 7
-  m_UserSelectedRegistryName: 
-  m_UserAddingNewScopedRegistry: 0
-  m_RegistryInfoDraft:
-    m_Modified: 0
-    m_ErrorMessage: 
-    m_UserModificationsInstanceId: -830
-    m_OriginalInstanceId: -832
-  m_LoadAssets: 0
