--- ProjectSettings/Packages/com.unity.testtools.codecoverage/Settings.json
@@ -1,7 +1,0 @@
-{
-    "m_Name": "Settings",
-    "m_Path": "ProjectSettings/Packages/com.unity.testtools.codecoverage/Settings.json",
-    "m_Dictionary": {
-        "m_DictionaryValues": []
-    }
-}
