--- ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,0 @@
-m_EditorVersion: 2021.2.7f1
-m_EditorVersionWithRevision: 2021.2.7f1 (6bd9e232123f)
