--- README.md
@@ -1,8 +1,0 @@
-# HistoryNavigator
-Simple history navigation toolbar for Unity 2021.2+ (ie. back/forward internet browser style)
-
-<img src="https://github.com/kronnect/HistoryNavigator/blob/master/demo.gif?raw=true">
-
-# Usage
-Click in Scene View to display the Overlays dialog. Enable the "History Navigator" toolbar overlay.
-
