--- UserSettings/Search.settings
@@ -1,1 +1,0 @@
-{}
