--- demo.gif
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:9adcd753041db8fdbc0b4f47f579277ee958d7921b145da7cace6250ef49b492
-size 6077709
