+++ package.json
@@ -1,0 +1,31 @@
+{
+  "name": "com.kronnect.tools.history-navigator",
+  "displayName": "Kronnect.Tools.HistoryNavigator",
+  "version": "0.0.0",
+  "unity": "2021.x",
+  "description": "Simple history navigation toolbar (ie. back/forward internet browser style)",
+  "keywords": [
+    "unity",
+	"unity3d",
+	"upm",
+    "common-games",
+	"kronnect",
+	"tools",
+    "utility"
+  ],
+  "homepage": "https://github.com/kronnect/HistoryNavigator",
+  "bugs": 
+  {
+    "url": "https://github.com/kronnect/HistoryNavigator/issues"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+ssh://git@github.com/kronnect/HistoryNavigator"
+  },
+  "license": "MIT",
+  "author": "Kronnect (edits by Common-Games)",
+  "scripts": 
+  {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  }
+}
